Move import to admin settings (fixes #361)
