+++ src/components/GrampsjsImport.js
@@ -1,12 +1,12 @@
-import {css, html} from 'lit'
-import '@material/mwc-select'
+import {css, html, LitElement} from 'lit'
 import '@material/mwc-button'
 
-import {GrampsjsView} from './GrampsjsView.js'
+import {sharedStyles} from '../SharedStyles.js'
+import {GrampsjsTranslateMixin} from '../mixins/GrampsjsTranslateMixin.js'
 import {apiPost} from '../api.js'
 import {fireEvent} from '../util.js'
-import '../components/GrampsjsFormUpload.js'
-import '../components/GrampsjsTaskProgressIndicator.js'
+import './GrampsjsFormUpload.js'
+import './GrampsjsTaskProgressIndicator.js'
 
 const STATE_ERROR = -1
 const STATE_INITIAL = 0
@@ -14,10 +14,10 @@ const STATE_READY = 1
 const STATE_PROGRESS = 2
 const STATE_DONE = 3
 
-export class GrampsjsViewImport extends GrampsjsView {
+export class GrampsjsImport extends GrampsjsTranslateMixin(LitElement) {
   static get styles() {
     return [
-      super.styles,
+      sharedStyles,
       css`
         .hidden {
           display: none;
@@ -40,13 +40,13 @@ export class GrampsjsViewImport extends GrampsjsView {
     this._uploadHint = ''
   }
 
-  renderContent() {
+  render() {
     return html`
-      <h2>${this._('Import')}</h2>
       <h3>${this._('Import Family Tree')}</h3>
 
       <p>
         <grampsjs-form-upload
+          outlined
           id="upload-tree"
           .strings="${this.strings}"
           filename
@@ -73,40 +73,6 @@ export class GrampsjsViewImport extends GrampsjsView {
           @task:error="${() => this._handleCompleted(STATE_ERROR)}"
         ></grampsjs-task-progress-indicator>
       </p>
-
-      <h3>${this._('Import Media Files')}</h3>
-      <p>
-        ${this._('Upload a ZIP archive with files for existing media objects.')}
-      </p>
-
-      <p>
-        <grampsjs-form-upload
-          accept="application/zip"
-          id="upload-media"
-          .strings="${this.strings}"
-          filename
-          @formdata:changed="${this._handleUploadChangedMedia}"
-        ></grampsjs-form-upload>
-      </p>
-      <p>
-        <mwc-button
-          raised
-          label="${this._('Import')}"
-          type="submit"
-          @click="${this._submitMedia}"
-          ?disabled=${this._mediaState !== STATE_READY}
-        ></mwc-button>
-        <grampsjs-task-progress-indicator
-          id="progress-media"
-          ?open="${this._mediaState !== STATE_INITIAL &&
-          this._mediaState !== STATE_READY}"
-          class="button"
-          size="20"
-          hideAfter="0"
-          @task:complete="${this._handleSuccessMedia}"
-          @task:error="${() => this._handleCompletedMedia(STATE_ERROR)}"
-        ></grampsjs-task-progress-indicator>
-      </p>
     `
   }
 
@@ -115,13 +81,6 @@ export class GrampsjsViewImport extends GrampsjsView {
       const uploadForm = this.shadowRoot.querySelector('#upload-tree')
       const ext = uploadForm.file.name.split('.').pop().toLowerCase()
       await this._submitTree(ext, uploadForm.file)
-    }
-  }
-
-  async _submitMedia() {
-    if (this._mediaState === STATE_READY) {
-      const uploadForm = this.shadowRoot.querySelector('#upload-media')
-      await this._submitMediaArchive(uploadForm.file)
     }
   }
 
@@ -144,37 +103,9 @@ export class GrampsjsViewImport extends GrampsjsView {
     }
   }
 
-  async _submitMediaArchive(file) {
-    this._mediaState = STATE_PROGRESS
-    const prog = this.renderRoot.querySelector('#progress-media')
-    prog.reset()
-    prog.open = true
-
-    const res = await apiPost(
-      '/api/media/archive/upload/zip',
-      file,
-      false,
-      false
-    )
-    if ('error' in res) {
-      prog.setError()
-      prog.errorMessage = this._(res.error)
-      this._handleCompletedMedia(STATE_ERROR)
-    } else if ('task' in res) {
-      prog.taskId = res.task?.id || ''
-    } else {
-      prog.setComplete()
-      this._handleSuccessMedia()
-    }
-  }
-
   _handleSuccess() {
     this._handleCompleted(STATE_DONE)
     fireEvent(this, 'db:changed', {})
-  }
-
-  _handleSuccessMedia() {
-    this._handleCompletedMedia(STATE_DONE)
   }
 
   _handleCompleted(state) {
@@ -182,12 +113,6 @@ export class GrampsjsViewImport extends GrampsjsView {
     const uploadForm = this.shadowRoot.querySelector('#upload-tree')
     uploadForm.reset()
     this._uploadHint = ''
-  }
-
-  _handleCompletedMedia(state) {
-    this._mediaState = state
-    const uploadForm = this.shadowRoot.querySelector('#upload-media')
-    uploadForm.reset()
   }
 
   _handleUploadChanged() {
@@ -229,15 +154,6 @@ export class GrampsjsViewImport extends GrampsjsView {
     }
     this._state = STATE_READY
   }
-
-  _handleUploadChangedMedia() {
-    const uploadForm = this.shadowRoot.querySelector('#upload-media')
-    if (!uploadForm.file?.name) {
-      this._mediaState = STATE_INITIAL
-      return
-    }
-    this._mediaState = STATE_READY
-  }
 }
 
-window.customElements.define('grampsjs-view-import', GrampsjsViewImport)
+window.customElements.define('grampsjs-import', GrampsjsImport)
