+++ src/views/GrampsjsViewSettings.js
@@ -1,19 +1,19 @@
 import {css, html} from 'lit'
 
 import {GrampsjsViewSettingsOnboarding} from './GrampsjsViewSettingsOnboarding.js'
+import './GrampsjsViewAdminSettings.js'
+import './GrampsjsViewUserManagement.js'
 import '../components/GrampsjsUsers.js'
 import '../components/GrampsjsTaskProgressIndicator.js'
 import '../components/GrampsjsShareUrl.js'
 import '../components/GrampsjsSysinfo.js'
 import '../components/GrampsjsTreeQuotas.js'
-import {doLogout, apiPost, apiPut, apiGet, getTreeId} from '../api.js'
+import {doLogout, apiPost, apiPut, getTreeId} from '../api.js'
 import '@material/mwc-textfield'
 import '@material/mwc-button'
 import '@material/mwc-select'
 import '@material/mwc-tab'
 import '@material/mwc-tab-bar'
-
-import {clickKeyHandler} from '../util.js'
 
 function renderLogoutButton() {
   return html`
@@ -32,27 +32,8 @@ export class GrampsjsViewSettings extends GrampsjsViewSettingsOnboarding {
     return [
       super.styles,
       css`
-        :host {
-        }
-
         mwc-tab-bar {
           margin-bottom: 30px;
-        }
-
-        grampsjs-share-url {
-          --mdc-icon-size: 18px;
-          --mdc-icon-button-size: 32px;
-          position: relative;
-          top: -5px;
-          color: rgba(0, 0, 0, 0.4);
-        }
-
-        span.url {
-          border-radius: 5px;
-          border: 1px solid #ddd;
-          font-size: 13px;
-          padding: 6px 6px;
-          color: #444;
         }
       `,
     ]
@@ -60,8 +41,7 @@ export class GrampsjsViewSettings extends GrampsjsViewSettingsOnboarding {
 
   static get properties() {
     return {
-      users: {type: Boolean},
-      userData: {type: Array},
+      owner: {type: Boolean},
       page: {type: String},
       dbInfo: {type: Object},
     }
@@ -69,8 +49,7 @@ export class GrampsjsViewSettings extends GrampsjsViewSettingsOnboarding {
 
   constructor() {
     super()
-    this.users = false
-    this.userData = []
+    this.owner = false
     this.page = 'user'
     this.dbInfo = {}
   }
@@ -85,13 +64,20 @@ export class GrampsjsViewSettings extends GrampsjsViewSettingsOnboarding {
             this.page = 'user'
           }}
         ></mwc-tab>
-        ${this.users
+        ${this.owner
           ? html`
               <mwc-tab
                 label="${this._('Administration')}"
                 ?active="${this.page === 'admin'}"
                 @click=${() => {
                   this.page = 'admin'
+                }}
+              ></mwc-tab>
+              <mwc-tab
+                label="${this._('Manage users')}"
+                ?active="${this.page === 'users'}"
+                @click=${() => {
+                  this.page = 'users'
                 }}
               ></mwc-tab>
             `
@@ -105,7 +91,8 @@ export class GrampsjsViewSettings extends GrampsjsViewSettingsOnboarding {
         ></mwc-tab>
       </mwc-tab-bar>
       ${this.page === 'user' ? this.renderUserSettings() : ''}
-      ${this.page === 'admin' && this.users ? this.renderAdminSettings() : ''}
+      ${this.page === 'admin' && this.owner ? this.renderAdminSettings() : ''}
+      ${this.page === 'users' && this.owner ? this.renderUserManagement() : ''}
       ${this.page === 'info' ? this.renderSysInfo() : ''}
     `
   }
@@ -119,56 +106,23 @@ export class GrampsjsViewSettings extends GrampsjsViewSettingsOnboarding {
 
   renderAdminSettings() {
     return html`
-      ${this.users
-        ? html`
-            <h3>${this._('Usage quotas')}</h3>
+      <grampsjs-view-admin-settings
+        active
+        .strings="${this.strings}"
+        .dbInfo="${this.dbInfo}"
+      >
+      </grampsjs-view-admin-settings>
+    `
+  }
 
-            <grampsjs-tree-quotas
-              .strings="${this.strings}"
-            ></grampsjs-tree-quotas>
-
-            <h3>${this._('Manage search index')}</h3>
-
-            <p>
-              ${this._(
-                'Manually updating the search index is usually unnecessary, but it may become necessary after an upgrade.'
-              )}
-            </p>
-            <mwc-button
-              outlined
-              @click="${this._updateSearch}"
-              @keydown="${clickKeyHandler}"
-              >${this._('Update search index')}</mwc-button
-            >
-            <grampsjs-task-progress-indicator
-              class="button"
-              id="progress-update-search"
-              size="20"
-            ></grampsjs-task-progress-indicator>
-
-            <h3>${this._('Manage users')}</h3>
-
-            ${this.dbInfo?.server?.multi_tree
-              ? html` <p>
-                  ${this._('Registration link')}:
-                  <span class="url">${this._registerUrl}</span>
-                  <grampsjs-share-url
-                    href="${this._registerUrl}"
-                  ></grampsjs-share-url>
-                </p>`
-              : ''}
-
-            <grampsjs-users
-              .strings="${this.strings}"
-              .data="${this.userData}"
-              ?ismulti="${!!this.dbInfo?.server?.multi_tree}"
-              @user:updated="${this._handleUserChanged}"
-              @user:added="${this._handleUserAdded}"
-              @user:added-multiple="${this._handleUsersAdded}"
-            >
-            </grampsjs-users>
-          `
-        : ''}
+  renderUserManagement() {
+    return html`
+      <grampsjs-view-user-management
+        active
+        .strings="${this.strings}"
+        .dbInfo="${this.dbInfo}"
+      >
+      </grampsjs-view-user-management>
     `
   }
 
@@ -247,80 +201,6 @@ export class GrampsjsViewSettings extends GrampsjsViewSettingsOnboarding {
     `
   }
 
-  firstUpdated() {
-    if (this.users) {
-      this._fetchUserData()
-    }
-  }
-
-  _handleUserChanged(e) {
-    const data = e.detail
-    this._updateUser(e.detail.name, {
-      role: data.role,
-      email: data.email,
-      full_name: data.full_name,
-    })
-  }
-
-  _handleUserAdded(e) {
-    const data = e.detail
-    this._addUser(e.detail.name, {
-      role: data.role,
-      email: data.email,
-      full_name: data.full_name,
-      password: data.password,
-    })
-  }
-
-  async _handleUsersAdded(e) {
-    const res = await apiPost('/api/users/', e.detail)
-    if ('error' in res) {
-      this.error = true
-      this._errorMessage = res.error
-    } else {
-      this.error = false
-      this._fetchUserData()
-    }
-  }
-
-  _updateUser(username, payload) {
-    apiPut(`/api/users/${username}/`, payload).then(data => {
-      if ('error' in data) {
-        this.error = true
-        this._errorMessage = data.error
-      } else {
-        this.error = false
-        this._fetchUserData()
-      }
-    })
-  }
-
-  _addUser(username, payload) {
-    apiPost(`/api/users/${username}/`, payload).then(data => {
-      if ('error' in data) {
-        this.error = true
-        this._errorMessage = data.error
-      } else {
-        this.error = false
-        this._fetchUserData()
-      }
-    })
-  }
-
-  _fetchUserData() {
-    this.loading = true
-    apiGet('/api/users/').then(data => {
-      if ('data' in data) {
-        this.error = false
-        this.userData = data.data
-      } else if ('error' in data) {
-        this.error = true
-        this._errorMessage = data.error
-      }
-      this.loading = false
-    })
-  }
-
   _changeEmail() {
     const form = this.shadowRoot.getElementById('change-email')
     if (!form.value) {
@@ -376,21 +256,6 @@ export class GrampsjsViewSettings extends GrampsjsViewSettingsOnboarding {
         formNewPw.value = ''
       }
     })
-  }
-
-  async _updateSearch() {
-    const prog = this.renderRoot.querySelector('#progress-update-search')
-    prog.reset()
-    prog.open = true
-    const data = await apiPost('/api/search/index/?full=1')
-    if ('error' in data) {
-      prog.setError()
-      prog.errorMessage = data.error
-    } else if ('task' in data) {
-      prog.taskId = data.task?.id || ''
-    } else {
-      prog.setComplete()
-    }
   }
 }
